fix: extract type declarations to app layer
